+++ src/index.ts
@@ -4,7 +4,7 @@
 
 'use strict';
 
-import ICalCalendar, {ICalCalendarData} from './calendar';
+import ICalCalendar, {ICalCalendarData} from './calendar.js';
 
 
 /**
@@ -14,7 +14,7 @@ import ICalCalendar, {ICalCalendarData} from './calendar';
  * import ical from 'ical-generator';
  *
  * // or use require:
- * // const ical = require('ical-generator');
+ * // const { default: ical } = require('ical-generator');
  *
  * const cal = ical();
  * ```
@@ -25,7 +25,7 @@ import ICalCalendar, {ICalCalendarData} from './calendar';
  * import ical from 'ical-generator';
  *
  * // or use require:
- * // const ical = require('ical-generator');
+ * // const { default: ical } = require('ical-generator');
  * const cal = ical({domain: 'sebbo.net'});
  *
  * // is the same as
@@ -53,7 +53,7 @@ export {
     ICalAlarmTypeValue,
     ICalAlarmJSONData,
     ICalAttachment
-} from './alarm';
+} from './alarm.js';
 
 export {
     default as ICalAttendee,
@@ -62,7 +62,7 @@ export {
     ICalAttendeeRole,
     ICalAttendeeStatus,
     ICalAttendeeJSONData
-} from './attendee';
+} from './attendee.js';
 
 export {
     default as ICalCalendar,
@@ -70,12 +70,12 @@ export {
     ICalCalendarProdIdData,
     ICalCalendarMethod,
     ICalCalendarJSONData
-} from './calendar';
+} from './calendar.js';
 
 export {
     default as ICalCategory,
     ICalCategoryData
-} from './category';
+} from './category.js';
 
 export {
     default as ICalEvent,
@@ -85,7 +85,7 @@ export {
     ICalEventData,
     ICalEventJSONData,
     ICalEventClass,
-} from './event';
+} from './event.js';
 
 export {
     ICalDateTimeValue,
@@ -103,16 +103,11 @@ export {
     ICalLuxonDateTimeStub,
     ICalDayJsStub,
     ICalRRuleStub
-} from './types';
+} from './types.js';
 
 export {
     formatDate,
     formatDateTZ,
     escape,
     foldLines
-} from './tools';
-
-/* istanbul ignore else */
-if (typeof module !== 'undefined') {
-    module.exports = Object.assign(ical, module.exports);
-}
+} from './tools.js';